tests/wpt/tests/fedcm/fedcm-authz/fedcm-continue-on-disallowed.https.html

<!DOCTYPE html>
<title>Federated Credential Management API network request tests.</title>
<link rel="help" href="https://fedidcg.github.io/FedCM">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-vendor.js"></script>

<body>

<script type="module">
import {fedcm_test,
        request_options_with_mediation_required,
        select_manifest,
        fedcm_get_and_select_first_account} from '../support/fedcm-helper.sub.js';

fedcm_test(async t => {
  // First, do a regular fedcm request so we that we can be considered
  // a returning user below.
  let options = request_options_with_mediation_required();
  await fedcm_get_and_select_first_account(t, options);

  // Now do a silent mediation request.
  options = request_options_with_mediation_required('manifest_with_continue_on.json');
  options.mediation = 'silent';
  await select_manifest(t, options);
  const cred_promise = fedcm_get_and_select_first_account(t, options);
  return promise_rejects_dom(t, 'NetworkError', cred_promise);
}, "continue_on with mediation:silent should fail");

</script>
